*** empty log message ***

+++ bundles/org.eclipse.swt/Eclipse SWT PI/gtk/org/eclipse/swt/internal/gtk/GdkRectangle.java
@@ -14,9 +14,9 @@ package org.eclipse.swt.internal.gtk;
  */
 
 public class GdkRectangle {
-	public short x;
-	public short y;
-	public short width;
-	public short height;
+	public int x;
+	public int y;
+	public int width;
+	public int height;
 	public static final int sizeof = 8;
 }
